Match the types with the tests with/without string promotion

+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/analysis/TypeCoercionSuite.scala
@@ -432,7 +432,7 @@ class TypeCoercionSuite extends PlanTest {
     widenTestWithoutStringPromotion(ArrayType(StringType), ArrayType(TimestampType), None)
 
     // String promotion
-    widenTestWithStringPromotion(LongType, StringType, Some(StringType))
+    widenTestWithStringPromotion(IntegerType, StringType, Some(StringType))
     widenTestWithStringPromotion(StringType, TimestampType, Some(StringType))
     widenTestWithStringPromotion(
       ArrayType(LongType), ArrayType(StringType), Some(ArrayType(StringType)))
